# Conflicts:
#	VerticalWarehouses/Machine Automation Service/MAS_FiniteStateMachines/FiniteStateMachines_Methods.cs
#	VerticalWarehouses/Machine Automation Service/MAS_FiniteStateMachines/Positioning/PositioningStateMachine.cs
#	VerticalWarehouses/Machine Automation Service/MAS_FiniteStateMachines/ShutterPositioning/ShutterPositioningStateMachine.cs

+++ VerticalWarehouses/Machine Automation Service/MAS_DataLayer/DataLayerService.BeltBurnishing.cs
@@ -9,7 +9,7 @@ namespace Ferretto.VW.MAS.DataLayer
         #region Properties
 
         /// <inheritdoc/>
-        public int CycleQuantity => this.GetIntegerConfigurationValue((long)BeltBurnishing.CycleQuantity, (long)ConfigurationCategory.BeltBurnishing);
+        public int CycleQuantity => this.GetIntegerConfigurationValue((long)BeltBurnishing.CycleQuantity, ConfigurationCategory.BeltBurnishing);
 
         #endregion
     }
