+++ VerticalWarehouses/Machine Automation Service/MAS_DataLayer/DataLayerService.ShutterHeightControl.cs
@@ -9,9 +9,9 @@ namespace Ferretto.VW.MAS.DataLayer
     {
         #region Properties
 
-        public decimal FeedRateSH => this.GetDecimalConfigurationValue((long)ShutterHeightControl.FeedRate, (long)ConfigurationCategory.ShutterHeightControl);
+        public decimal FeedRateSH => this.GetDecimalConfigurationValue((long)ShutterHeightControl.FeedRate, ConfigurationCategory.ShutterHeightControl);
 
-        public decimal RequiredTolerance => this.GetDecimalConfigurationValue((long)ShutterHeightControl.RequiredTolerance, (long)ConfigurationCategory.ShutterHeightControl);
+        public decimal RequiredTolerance => this.GetDecimalConfigurationValue((long)ShutterHeightControl.RequiredTolerance, ConfigurationCategory.ShutterHeightControl);
 
         #endregion
     }
